% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{orf_plot}
\alias{orf_plot}
\title{Plot ORFs (Open Reading Frames) for Each Transcript}
\usage{
orf_plot(data = NULL, trans_col = "#996600")
}
\arguments{
\item{data}{A data frame containing information about the ORFs. It must contain the columns \code{start}, \code{end}, \code{sense}, \code{orf}, and \code{id} or alternatively the columns \code{st_start} and \code{sp_stop}.}

\item{trans_col}{A color string representing the color for the transcript line in the plot. Default is \verb{#996600}.}
}
\value{
A combined ggplot object using \code{cowplot::plot_grid} to visualize the ORFs for all transcripts.
The resulting plot will display individual ORFs as colored rectangles along the transcript axis, with different colors for \code{+} and \code{-} strands.
}
\description{
This function generates visualizations of Open Reading Frames (ORFs) across transcript sequences using \code{ggplot2}.
For each transcript (specified by the "id" column), ORFs are shown as colored rectangles based on their strand direction ('+' or '-').
A separate plot is created for each transcript, and all plots are combined together.
}
